KafkaListener/modules/KafkaModule/requirements.txt
azure-iot-device~=2.7.0
confluent-kafka~=1.7.0
aiokafka~=0.7.2
